refactor: 컴포넌트 스타일 상수 분리
